Making the script idempotent

+++ vcx/libvcx/build_scripts/ios/mac/mac.norms.local.pod.install.sh
@@ -15,10 +15,11 @@ COCOAPOD_SPEC=/Users/norm/.cocoapods/repos/evernym-1/Specs/vcx/0.0.15/vcx.podspe
 
 cd ${COCOAPOD_BUILD_FOLDER}
 COCOAPOD=$(ls *.zip)
-rm ${PODSERVER_DOCROOT}/*.zip
-mv "${COCOAPOD_BUILD_FOLDER}/${COCOAPOD}" ${PODSERVER_DOCROOT}
-
-sed -i .bak "s/vcx\.framework_[0-9]*\.[0-9]*_universal\.zip/${COCOAPOD}/" ${COCOAPOD_SPEC}
+if [ -f "${COCOAPOD_BUILD_FOLDER}/${COCOAPOD}" ]; then
+    rm ${PODSERVER_DOCROOT}/*.zip
+    mv "${COCOAPOD_BUILD_FOLDER}/${COCOAPOD}" ${PODSERVER_DOCROOT}
+    sed -i .bak "s/vcx\.framework_[0-9]*\.[0-9]*_universal\.zip/${COCOAPOD}/" ${COCOAPOD_SPEC}
+fi
 
 #COCOAPOD=kdkdkdkd
 #sed -i .bak "s/vcx\.framework_[0-9]*\.[0-9]*_universal\.zip/${COCOAPOD}/" /Users/norm/.cocoapods/repos/evernym-1/Specs/vcx/0.0.15/vcx.podspec
